[Changed] Add/init webSockPort & httpJSONPort in net/node By the way, fix httpInfoPort from int to uint16

+++ test/monitor/ngbnodeinfo.go
@@ -7,7 +7,7 @@ type NgbNodeInfo struct {
 	NgbType       string
 	NgbAddr       string
 	HttpInfoAddr  string
-	HttpInfoPort  int
+	HttpInfoPort  uint16
 	HttpInfoStart bool
 }
